# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Landing Lens API",
        "description": "A powerful REST API for AI-powered image object detection using LandingLens technology.",
        "contact": {
            "name": "API Support",
            "email": "support@landinglens-api.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://landingapi.test/",
            "description": "Development Server"
        },
        {
            "url": "https://your-domain.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/predict": {
            "get": {
                "tags": [
                    "Image Analysis"
                ],
                "summary": "Get paginated list of recent analyses",
                "description": "List recent analyses",
                "operationId": "9daf083f476fbdd6cf86279c952af0f3",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1,
                            "example": 10
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "processing",
                                "completed",
                                "failed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recent analyses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "analysis_id": {
                                                        "type": "integer",
                                                        "example": 123
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "completed"
                                                    },
                                                    "objects_detected_count": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "average_confidence": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 0.86999999999999999555910790149937383830547332763671875
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 47
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "has_more": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Image Analysis"
                ],
                "summary": "Analyze image for object detection using AI",
                "description": "Analyze image for object detection",
                "operationId": "930f3494a35e688f9f0f4532bc384577",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Image file (JPEG, PNG, JPG, max 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "callback_url": {
                                        "description": "Optional webhook URL for completion notification",
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/webhook"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Image analyzed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Image analyzed successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ImageAnalysis"
                                        },
                                        "links": {
                                            "properties": {
                                                "self": {
                                                    "type": "string",
                                                    "format": "url"
                                                },
                                                "original_image": {
                                                    "type": "string",
                                                    "format": "url"
                                                },
                                                "processed_image": {
                                                    "type": "string",
                                                    "format": "url"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Processing error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ImageAnalysis": {
                "properties": {
                    "analysis_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed"
                        ],
                        "example": "completed"
                    },
                    "original_filename": {
                        "type": "string",
                        "example": "image.jpg"
                    },
                    "file_size": {
                        "type": "integer",
                        "example": 245760
                    },
                    "mime_type": {
                        "type": "string",
                        "example": "image/jpeg"
                    },
                    "objects_detected_count": {
                        "type": "integer",
                        "example": 3
                    },
                    "average_confidence": {
                        "type": "number",
                        "format": "float",
                        "example": 0.86999999999999999555910790149937383830547332763671875
                    },
                    "processing_time_seconds": {
                        "type": "number",
                        "format": "float",
                        "example": 2.45000000000000017763568394002504646778106689453125
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "images": {
                        "properties": {
                            "original_url": {
                                "type": "string",
                                "format": "url"
                            },
                            "processed_url": {
                                "type": "string",
                                "format": "url"
                            }
                        },
                        "type": "object"
                    },
                    "detected_objects": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DetectedObject"
                        }
                    },
                    "api_response": {
                        "properties": {
                            "model_id": {
                                "type": "string",
                                "example": "model_123"
                            },
                            "backbone_type": {
                                "type": "string",
                                "example": "efficientnet"
                            },
                            "latency_seconds": {
                                "type": "number",
                                "format": "float",
                                "example": 2.45000000000000017763568394002504646778106689453125
                            }
                        },
                        "type": "object"
                    },
                    "error_message": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}